api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aniso8601==4.1.0          # via flask-restful
ciso8601==2.1.1
click==7.0                # via flask
cytoolz==0.9.0.1
flask-restful==0.3.7
flask==1.0.2
itsdangerous==1.1.0       # via flask
jinja2==2.10              # via flask
markupsafe==1.1.0         # via jinja2
peewee==3.8.2
psycopg2-binary==2.7.6.1
python-dateutil==2.7.5
pytz==2018.9              # via flask-restful
pyyaml==4.2b4
six==1.12.0               # via flask-restful, python-dateutil
toolz==0.9.0              # via cytoolz
werkzeug==0.14.1          # via flask
